fix: change images type and return file names in store/types
