+++ apps/client/src/__mocks__/store/state.ts
@@ -1,13 +1,11 @@
-import { type State } from '@/store/types/state.types';
+import { type State } from '@/store/types/state';
 import { Language } from '@types';
 
 const mockInitialState: State = {
   id: null,
   username: null,
   email: null,
-  password: null,
   authStatus: false,
-  avatarUrl: null,
   language: Language.RU,
   registration: { fields: {}, isFormValid: false },
   login: { fields: {}, isFormValid: false },
