+++ apps/client/src/pages/WelcomePage/WelcomePage.ts
@@ -6,7 +6,7 @@ import type { Action } from '@/api/StateAPI';
 import store from '@store/store';
 import RegButton from './RegButton/RegButton';
 import { WelcomeActions } from '@/store/actions/welcome.actions';
-import type { State } from '@store/types/state.types';
+import type { State } from '@/store/types/state';
 import { TestActions } from '@/store/actions/test.actions';
 import LangButton from './LangButton/LangButton';
