+++ apps/client/src/store/middlewares/test.sender.middleware.ts
@@ -1,5 +1,5 @@
 import type { Middleware } from '@/api/StateAPI/types/types';
-import type { AppActions } from '@store/types/action.types';
+import type { AppActions } from '@/store/types/action';
 import { TestActions } from '../actions/test.actions';
 
 export default function sender<State>(): Middleware<State, AppActions> {
