+++ apps/client/src/store/types/state.ts
@@ -1,0 +1,13 @@
+import type { FormState } from '@/components/BaseForm/BaseFormTypes';
+import { Language } from '@/types';
+
+export type State = {
+  id: string | null;
+  username: string | null;
+  email: string | null;
+  authStatus: boolean;
+  language: Language;
+  registration: FormState;
+  login: FormState;
+  profile: FormState;
+};
