--- apps/client/src/store/types/state.types.ts
@@ -1,15 +1,0 @@
-import type { FormState } from '@/components/BaseForm/BaseFormTypes';
-import { Language } from '@/types';
-
-export type State = {
-  id: string | null;
-  username: string | null;
-  password: string | null;
-  email: string | null;
-  avatarUrl: string | null;
-  authStatus: boolean;
-  language: Language;
-  registration: FormState;
-  login: FormState;
-  profile: FormState;
-};
